Move a little bit of code around; in particular, all file interface subclasses can share functions since they already read in pandas.
